录制方向，RSA，横竖屏
1.录屏横竖屏和锁定适配
2.RSA
3.日期显示的bug
4.其他部分优化

+++ Classes/Core/FanAudioConverter.h
@@ -58,7 +58,7 @@ typedef NS_ENUM(NSUInteger, FanAudioConverterErrorType) {
 - (void)fan_audioConverter:(FanAudioConverter*)converter nextBytes:(char*)bytes length:(NSUInteger*)length;
 @optional
 ///转换偏移量到多少字节  
-- (void)fan_audioConverter:(FanAudioConverter *)converter seekToPosition:(NSUInteger)position;
+- (void)fan_audioConverter:(FanAudioConverter *)converter seekToPosition:(long long)position;
 @end
 
 #pragma mark -FanAudioConverter
